+++ projects/angular-cesium/src/lib/angular-cesium/models/map-layer-provider-options.enum.ts
@@ -11,4 +11,3 @@ export enum MapLayerProviderOptions {
   UrlTemplateImagery = Cesium.UrlTemplateImageryProvider,
   OFFLINE = null,
 }
-