+++ projects/angular-cesium/src/lib/angular-cesium/models/map-terrain-provider-options.enum.ts
@@ -1,0 +1,8 @@
+export enum MapTerrainProviderOptions {
+  CesiumTerrain = Cesium.CesiumTerrainProvider,
+  ArcGISTiledElevation = Cesium.ArcGISTiledElevationTerrainProvider,
+  GoogleEarthEnterprise = Cesium.GoogleEarthEnterpriseTerrainProvider,
+  VRTheWorld = Cesium.VRTheWorldTerrainProvider,
+  Ellipsoid = Cesium.EllipsoidTerrainProvider,
+  WorldTerrain = Cesium.createWorldTerrain
+}