+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 ### Features
 * Added a rectangle drawing widget
 * Added an example rectangle component in the demo app
+* Added `ac-map-terrain-provider` component which supports Cesium terrain providers
+* Added `ac-map-terrain-provider` component to demo app to show it's usage
 
 ## 0.59
 ### Fixes 
@@ -13,7 +15,7 @@
 * Fix `AcHtmlDesc` remove.
 * Add doc to the `ContextMenuService`
 ### Features
-* Add support for terrian in Polygon and Polyline Editors.
+* Add support for terrain in Polygon and Polyline Editors.
 * To enable terrain support: `this.polylineEditor.create({ clampHeightTo3D: true, clampHeightTo3DOptions: {clampToTerrain: true} });` 
 
 ### Breaking Changes